+++ tests/testthat/test-gov_button.R
@@ -6,18 +6,18 @@ test_that("button works", {
   button <- button_Input("btn1", "Click me")
 
   expect_identical(
-    button$attribs$id,
+    htmltools::tagGetAttribute(button, "id"),
     "btn1"
   )
 
   expect_identical(
-    button$attribs$class,
+    htmltools::tagGetAttribute(button, "class"),
     "govuk-button action-button"
   )
 
   expect_identical(
     "Click me",
-    button$children[[1]]
+    tag_text(button, "govuk-button")
   )
 })
 
@@ -25,18 +25,20 @@ test_that("start_button works", {
   button <- button_Input("btn1", "Click me", type = "start")
 
   expect_identical(
-    button$attribs$id,
+    htmltools::tagGetAttribute(button, "id"),
     "btn1"
   )
 
   expect_identical(
-    button$attribs$class,
+    htmltools::tagGetAttribute(button, "class"),
     "govuk-button govuk-button--start action-button"
   )
 
+  # A start button carries the label plus the arrow svg, so assert the label
+  # child rather than reaching for the single-child tag_text().
   expect_identical(
     "Click me",
-    button$children[[1]]
+    rendered_children(button)[[1L]]
   )
 })
 
@@ -44,18 +46,18 @@ test_that("secondary_button works", {
   button <- button_Input("btn1", "Click me", type = "secondary")
 
   expect_identical(
-    button$attribs$id,
+    htmltools::tagGetAttribute(button, "id"),
     "btn1"
   )
 
   expect_identical(
-    button$attribs$class,
+    htmltools::tagGetAttribute(button, "class"),
     "govuk-button govuk-button--secondary action-button"
   )
 
   expect_identical(
     "Click me",
-    button$children[[1]]
+    tag_text(button, "govuk-button")
   )
 })
 
@@ -63,17 +65,17 @@ test_that("warning_button works", {
   button <- button_Input("btn1", "Click me", type = "warning")
 
   expect_identical(
-    button$attribs$id,
+    htmltools::tagGetAttribute(button, "id"),
     "btn1"
   )
 
   expect_identical(
-    button$attribs$class,
+    htmltools::tagGetAttribute(button, "class"),
     "govuk-button govuk-button--warning action-button"
   )
 
   expect_identical(
     "Click me",
-    button$children[[1]]
+    tag_text(button, "govuk-button")
   )
 })