+++ tests/testthat/test-heading_text.R
@@ -2,13 +2,13 @@ test_that("default", {
   head_text <- heading_text("Test Time")
 
   expect_identical(
-    head_text$attribs$class,
+    htmltools::tagGetAttribute(head_text, "class"),
     "govuk-heading-xl"
   )
 
   expect_identical(
     shiny::HTML("Test Time"),
-    head_text$children[[1]]
+    tag_text(head_text, "govuk-heading-xl")
   )
 })
 
@@ -16,12 +16,12 @@ test_that("medium_works", {
   head_text <- heading_text("Test Time", "m")
 
   expect_identical(
-    head_text$attribs$class,
+    htmltools::tagGetAttribute(head_text, "class"),
     "govuk-heading-m"
   )
 
   expect_identical(
     shiny::HTML("Test Time"),
-    head_text$children[[1]]
+    tag_text(head_text, "govuk-heading-m")
   )
 })